backend/pyproject.toml
[project]
name = "movie-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.1",
    "aiosqlite>=0.22.1",
    "alembic>=1.18.3",
    "cachetools>=5.5.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.40.0",
    "python-jose[cryptography]>=3.3.0",
]
